# prescription-service/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for managing patient prescriptions, including creation, retrieval, and status management",
        "title": "Prescription Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "PulseOne Support",
            "url": "http://www.pulseone.com/support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8085",
    "basePath": "/api",
    "paths": {
        "/prescriptions": {
            "get": {
                "description": "Retrieve prescriptions for a clinic with optional filters. Used by clinic admins to see pending prescriptions.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prescriptions"
                ],
                "summary": "Get prescriptions for clinic dashboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clinic ID",
                        "name": "clinic_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "ACTIVE",
                        "description": "Status: ACTIVE, FILLED, CANCELLED",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date filter: TODAY or YYYY-MM-DD format",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Prescription"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new prescription with items (drugs). Default status is ACTIVE.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prescriptions"
                ],
                "summary": "Create a new prescription",
                "parameters": [
                    {
                        "description": "Prescription request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreatePrescriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Prescription"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/prescriptions/active": {
            "get": {
                "description": "Retrieve all active prescriptions system-wide, useful for pharmacy operations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prescriptions"
                ],
                "summary": "Get all active prescriptions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by clinic ID",
                        "name": "clinic_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Prescription"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/prescriptions/appointment/{appointment_id}": {
            "get": {
                "description": "Retrieve prescription details for a specific appointment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prescriptions"
                ],
                "summary": "Get prescription by appointment ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Appointment ID",
                        "name": "appointment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Prescription"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/prescriptions/doctor/{doctor_id}": {
            "get": {
                "description": "Retrieve all prescriptions issued by a specific doctor",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prescriptions"
                ],
                "summary": "Get prescriptions by doctor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Doctor ID",
                        "name": "doctor_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Status filter: ACTIVE, FILLED, CANCELLED",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Prescription"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/prescriptions/patient/{patient_id}": {
            "get": {
                "description": "Retrieve complete prescription history for a specific patient",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prescriptions"
                ],
                "summary": "Get patient prescription history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "patient_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Prescription"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/prescriptions/stats": {
            "get": {
                "description": "Get prescription statistics for a clinic or doctor",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prescriptions"
                ],
                "summary": "Get prescription statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clinic ID",
                        "name": "clinic_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Doctor ID",
                        "name": "doctor_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/prescriptions/{id}/status": {
            "patch": {
                "description": "Update prescription status (ACTIVE, FILLED, CANCELLED). Used when admin dispenses drugs or external pharmacy completes the order.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prescriptions"
                ],
                "summary": "Update prescription status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Prescription ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Prescription"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.CreatePrescriptionRequest": {
            "type": "object",
            "properties": {
                "appointment_id": {
                    "type": "string"
                },
                "clinic_id": {
                    "type": "string"
                },
                "doctor_id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CreatePrescriptionItemRequest"
                    }
                },
                "patient_id": {
                    "type": "string"
                }
            }
        },
        "model.Prescription": {
            "type": "object",
            "properties": {
                "appointment_id": {
                    "type": "string"
                },
                "clinic_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "doctor_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "issued_at": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PrescriptionItem"
                    }
                },
                "patient_id": {
                    "type": "string"
                },
                "status": {
                    "description": "ACTIVE, FILLED, CANCELLED",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateStatusRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        }
    }
}